requirements.txt
requests
openai
yfinance
pandas
numpy
scipy
python-dotenv
rich
